requirements.txt
fastapi
pydantic
pydantic[email]
PyMySQL
SQLAlchemy
uvicorn[standard]

# jwt auth
pyjwt
bcrypt

# test
pytest
